Corretto bug conversion factor

+++ frontend/src/app/sensor/gauge/pressure/pressureGauge.component.ts
@@ -67,7 +67,7 @@ export class PressureGaugeComponent implements OnInit, OnChanges {
 
   onChangeObj(event) {
 
-    if(this.isConversionfactor()){
+    if(this.isConversionfactor() && this.user != null){
       this.userService.removeUnitKnowledgeUser(this.user, this.conversionFactors[this.unitConverterSelected].id)
         .subscribe(user => {
           if (user){
@@ -80,7 +80,7 @@ export class PressureGaugeComponent implements OnInit, OnChanges {
 
     this.unitConverterSelected = event.value + 0;
 
-    if(this.isConversionfactor()){
+    if(this.isConversionfactor() && this.user != null){
       this.userService.addUnitKnowledgeUser(this.user, this.conversionFactors[this.unitConverterSelected].id)
         .subscribe(user => {
           if (user){
