+++ frontend/src/app/sensor/gauge/temperature/temperatureGauge.component.ts
@@ -68,7 +68,7 @@ export class TemperatureGaugeComponent implements OnInit, OnChanges {
 
   onChangeObj(event) {
 
-    if(this.isConversionfactor()){
+    if(this.isConversionfactor() && this.user != null){
       this.userService.removeUnitKnowledgeUser(this.user, this.conversionFactors[this.unitConverterSelected].id)
         .subscribe(user => {
           if (user){
@@ -81,7 +81,7 @@ export class TemperatureGaugeComponent implements OnInit, OnChanges {
 
     this.unitConverterSelected = event.value + 0;
 
-    if(this.isConversionfactor()){
+    if(this.isConversionfactor() && this.user != null){
       this.userService.addUnitKnowledgeUser(this.user, this.conversionFactors[this.unitConverterSelected].id)
         .subscribe(user => {
           if (user){
