+++ frontend/src/app/sensor/gauge/uv/uvGauge.component.ts
@@ -66,7 +66,7 @@ export class UVGaugeComponent implements OnInit, OnChanges {
 
   onChangeObj(event) {
 
-    if(this.isConversionfactor()){
+    if(this.isConversionfactor() && this.user != null){
       this.userService.removeUnitKnowledgeUser(this.user, this.conversionFactors[this.unitConverterSelected].id)
         .subscribe(user => {
           if (user){
@@ -79,7 +79,7 @@ export class UVGaugeComponent implements OnInit, OnChanges {
 
     this.unitConverterSelected = event.value + 0;
 
-    if(this.isConversionfactor()){
+    if(this.isConversionfactor() && this.user != null){
       this.userService.addUnitKnowledgeUser(this.user, this.conversionFactors[this.unitConverterSelected].id)
         .subscribe(user => {
           if (user){
